Removed use of base object DOMAIN constant and replaced with a string literal.
The aim of the constant was to minimize the chances of the 'namespace prefix' from being misspelled and/or to make it easy to update en masse in the future.

However outside of base class descendants a different mechanism had to be used anyway - and it was probably not l10n/i18n friendly.