Track shortcode usages/locations.
Some shortcodes are fundamental to a nice end user experience - the front editor and post list shortcodes for instance; there should be only one of each and they can be moved from page to page.

This piece of work allows for the location of those shortcodes to be detected and recalled with low overhead.